requirements.txt
Flask>=3.0
numpy>=1.24
Pillow>=10.0
pandas>=2.0
